Moves the sample job scheduler extension plugin to the plugins/examples directory
