+++ plugins/examples/job-scheduler-sample-extension/src/test/java/org/opensearch/example/jobschedulersampleextension/SampleExtensionPluginIT.java
@@ -1,9 +1,12 @@
 /*
- * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
  */
 
-package org.opensearch.jobscheduler.smapleextension;
+package org.opensearch.example.jobschedulersampleextension;
 
 import org.opensearch.action.admin.cluster.health.ClusterHealthRequest;
 import org.opensearch.action.admin.cluster.health.ClusterHealthResponse;
